Start to separate Civiimport parser functions from shared parent
